server/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
google-generativeai
httpx
pydantic-settings
python-multipart
sse-starlette
python-dotenv
reportlab
